Simplified nested outline syntax
Orderedness of a list is now controlled by its parent.  This prevents mixing of
unordered and ordered list elements in the same list.  The resulting clarity
paves the way for the full tree-based rewrite.

Also, this fixes a bug where an ordered list which starts with a comment would
be blocked.  Comments are now properly ignored for blocking purposes.