Be able to display multiline actions
The text for an action might have newlines in it.  Previously, this would have
caused subsequent actions to become misaligned (so interacting with an action on
one line would actually affect an action on another).  Now, a multiline action
is split into several lines, each of which points to the action node.