Update system before adding context flag callback
When maktaba adds a callback for a flag, it immediately runs the
callback once.  For some users, this could happen before the global
'my_system' python variable was defined.  Calling 'vtd#UpdateSystem()'
first, makes sure that 'my_system' is defined.

This is yet another indication that I need to refactor and simplify the
code surrounding the VTD view buffer, and its interaction with the
trusted system.  That gargantuan task is for another day.

Fixes #13.