Python should receive a raw string for sys.path
Python is sent a path to find the libvtd module. The call is done by
executing python code as a string directly. However, the string passed
is not escaped, and with Windows paths (containing backslashes) this
means Python converts all escape sequences.

Solved by adding an r before the string literal, making it into a raw
Python string, which does not escape the backslashes.

Fixes #21.