Use sorted views instead of sorting containers
`python3` made these things not be containers, so we can't call sort on
them anymore.

Unfortunately, for anything in `vim-vtd` (as opposed to `libvtd`), I'm
pretty much guessing as far as `python2` compatibility is concerned.  I
don't expect to go to the trouble to compile a special version of vim
with this compatibility built in.

With this commit, VTD now runs!

Helps #24.