Update special char byte-decoding for python3
The magic incantation is: a byte literal, decoded via UTF-8.  (In
python3, you don't need to specify which decoding to use because UTF-8
is the default, but it's a bad idea to rely on this in general, and in
this case it would break python2 support.)

After this diff, special characters for recurring and waiting actions
will show up properly on modern systems.

Helps #24.