Encode the key to the vim dictionary
This fixes the statusline, which was broken.  It was an extremely tricky
bug to track down.

I noticed that the counts on the statusline were wrong: specifically, I
soon realized that values were getting clobbered.

For example, suppose I had 4 Next Actions (`X`) that were due, and 1
Waiting (`D` for "delegated").

Expected:
  DUE: 4[X], 1[D]
Actual:
  DUE: 1[D]

Eventually, I tracked down the source of the problem to these lines:

    if key not in category_counts.keys():
        category_counts[key] = {}

Makes sense, since this _would_ do the clobbering... but only if the
`not in` check fails.  And if it fails, then shouldn't there be nothing
to clobber?!

It turns out that python3 changes the default string representation to
actual (unicode) strings... but, presumably, the keys in vim
dictionaries (such as `category_counts`) must be `bytes`, i.e., encoded
unicode strings [1].  Thus, `not in` will **always return `true`**, and
we always get clobbering.

The solution is to encode the key via UTF-8.  This is correct in python3
(I think), but probably technically wrong in python2.  However, it works
as long as the keys happen to be restricted to ASCII codes (0-127).  In
this case, I belive they are because
  a) if they weren't, presumably python2 would have mis-handled them;
     and,
  b) I have a fully enumerated list of these keys anyway (in
     `libvtd.node`).

OK, so now the statusline is correct, so as far as I can tell, vim-vtd
is working just as well with modern, python3-based versions as it did
with python2!

[1] https://nedbatchelder.com/text/unipain.html

Fixes #24.